+++ models/Individual.py
@@ -1,0 +1,11 @@
+class Individual:
+    __slots__ = "id", "name", "gender", "birthday", "age", "alive", "death", "child", "spouse"
+
+    def __init__(self, unique_id):
+        self.id = unique_id
+        self.child = {}
+        self.spouse = {}
+        self.age = ""
+        self.alive = True
+        self.death = "NA"
+        # TODO define age determinations based on birthday.
